Started laying down foundation for template
Sectioned out different pages of portfolio so far (About, Work)

The styling is made so that the pages are aligned side by side, and
held together by a container div. When clicking to go on another page,
the large div will slide itself to show the next page. (Still needs to
be implemented)

I used SASS to do the styling for the different sections and created
different css files for each important area for the portfolio so that
the file we work on will be separated and organized and not all in one
chaotic file. The way these different css are linked together is by
import calls made at the end of main.css.